add task solution 1
