+++ src/components/TodosList/TodosList.tsx
@@ -1,0 +1,80 @@
+import React from 'react';
+import { Todo } from '../../types/Todo';
+import classNames from 'classnames';
+
+interface Props {
+  todos: Todo[];
+  activeTodo: Todo | null;
+  onSelectTodo: (v: Todo) => void;
+}
+export const TodosList: React.FC<Props> = ({
+  todos,
+  activeTodo,
+  onSelectTodo,
+}) => {
+  return (
+    <table className="table is-narrow is-fullwidth">
+      <thead>
+        <tr>
+          <th>#</th>
+          <th>
+            <span className="icon">
+              <i className="fas fa-check" />
+            </span>
+          </th>
+          <th>Title</th>
+          <th> </th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {todos.map(todo => {
+          return (
+            <tr
+              key={todo.id}
+              data-cy="todo"
+              className={
+                activeTodo?.id === todo.id ? 'has-background-info-light' : ''
+              }
+            >
+              <td className="is-vcentered">{todo.id}</td>
+              <td className="is-vcentered">
+                {todo.completed && (
+                  <span className="icon" data-cy="iconCompleted">
+                    <i className="fas fa-check" />
+                  </span>
+                )}
+              </td>
+              <td className="is-vcentered is-expanded">
+                <p
+                  className={
+                    todo.completed ? 'has-text-success' : 'has-text-danger'
+                  }
+                >
+                  {todo.title}
+                </p>
+              </td>
+              <td className="has-text-right is-vcentered">
+                <button
+                  data-cy="selectButton"
+                  className="button"
+                  type="button"
+                  onClick={() => onSelectTodo(todo)}
+                >
+                  <span className="icon">
+                    <i
+                      className={classNames(
+                        'far',
+                        activeTodo?.id === todo.id ? 'fa-eye-slash' : 'fa-eye',
+                      )}
+                    />
+                  </span>
+                </button>
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
